use 64bit feature

+++ kernel/feature.h
@@ -10,11 +10,11 @@ enum ksu_feature_id {
     KSU_FEATURE_MAX
 };
 
-typedef int (*ksu_feature_get_t)(u32 *value);
-typedef int (*ksu_feature_set_t)(u32 value);
+typedef int (*ksu_feature_get_t)(u64 *value);
+typedef int (*ksu_feature_set_t)(u64 value);
 
 struct ksu_feature_handler {
-    u32 feature_id;
+    u64 feature_id;
     const char *name;
     ksu_feature_get_t get_handler;
     ksu_feature_set_t set_handler;
@@ -22,11 +22,11 @@ struct ksu_feature_handler {
 
 int ksu_register_feature_handler(const struct ksu_feature_handler *handler);
 
-int ksu_unregister_feature_handler(u32 feature_id);
+int ksu_unregister_feature_handler(u64 feature_id);
 
-int ksu_get_feature(u32 feature_id, u32 *value, bool *supported);
+int ksu_get_feature(u64 feature_id, u64 *value, bool *supported);
 
-int ksu_set_feature(u32 feature_id, u32 value);
+int ksu_set_feature(u64 feature_id, u64 value);
 
 void ksu_feature_init(void);
